requirements.txt
numpy
pandas 
scikit-learn
matplotlib
streamlit
plotly.express